Rename authorization_list_parser module
